pyproject.toml
[project]
name = "pycrusher"
version = "0.3.9"
description = "Generate lossy JPEG compressions for fun!"
readme = "README.md"
requires-python = ">=3.7"
dependencies = ["pillow>=9.5.0", "tqdm>=4.66.5"]

[tool.uv]
dev-dependencies = [
  "hypothesis>=6.79.4",
  "mypy>=1.4.1",
  "pytest>=7.4.4",
  "types-colorama>=0.4.15.12",
  "types-pillow>=10.1.0.2",
  "types-pygments>=2.17.0.0",
  "types-setuptools>=69.0.0.0",
  "types-tqdm>=4.66.0.5",
]
